Modularize exporter package
