+++ proxmox_exporter/__init__.py
@@ -1,0 +1,5 @@
+"""Proxmox exporter package."""
+
+from .exporter import EnhancedProxmoxExporter
+
+__all__ = ["EnhancedProxmoxExporter"]
